pkg/git: Move pkg/util/git.go functions into main git library
