+++ cmd/krel/cmd/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/command:go_default_library",
+        "//pkg/git:go_default_library",
         "//pkg/util:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
         "@in_gopkg_src_d_go_git_v4//:go_default_library",
