+++ pkg/git/BUILD.bazel
@@ -7,8 +7,12 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_blang_semver//:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
         "@in_gopkg_src_d_go_git_v4//:go_default_library",
+        "@in_gopkg_src_d_go_git_v4//config:go_default_library",
         "@in_gopkg_src_d_go_git_v4//plumbing:go_default_library",
+        "@in_gopkg_src_d_go_git_v4//plumbing/object:go_default_library",
+        "@in_gopkg_src_d_go_git_v4//storage/memory:go_default_library",
     ],
 )
