+++ pkg/util/BUILD.bazel
@@ -2,20 +2,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "common.go",
-        "git.go",
-    ],
+    srcs = ["common.go"],
     importpath = "k8s.io/release/pkg/util",
     visibility = ["//visibility:public"],
-    deps = [
-        "@com_github_pkg_errors//:go_default_library",
-        "@in_gopkg_src_d_go_git_v4//:go_default_library",
-        "@in_gopkg_src_d_go_git_v4//config:go_default_library",
-        "@in_gopkg_src_d_go_git_v4//plumbing:go_default_library",
-        "@in_gopkg_src_d_go_git_v4//plumbing/object:go_default_library",
-        "@in_gopkg_src_d_go_git_v4//storage/memory:go_default_library",
-    ],
 )
 
 filegroup(
